Move views and locale location
