+++ app/server/utils/fieldValidation.ts
@@ -1,5 +1,5 @@
 const Joi = require('joi');
-const i18n = require('../locale/en');
+const i18n = require('../../../locale/en');
 
 const maxCharacters = 10000;
